task/backend/requirements.txt
Django==5.2.1
# バージョン固定
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
drf-spectacular==0.27.1
gunicorn
